refactor(ui): unify badge family + fold modem-merge helper
